requirements.txt
websocket-client
websockets
python-dotenv
requests
rugcheck
httpx
